requirements.txt
datacite
langdetect
pandas
#datahugger
datahugger-ng
pyo3_runtime
